Finished parsing and fixed some bugs

+++ DataTypes/TimingPoint.cs
@@ -5,7 +5,13 @@ namespace OsuFormatReader.DataTypes;
 public class TimingPoint
 {
     public int time { get; set; }
-    public decimal beatLength  { get; set; }
+    /// <summary>
+    /// Wiki description: https://osu.ppy.sh/wiki/en/Client/File_formats/osu_%28file_format%29#hitsounds:~:text=beatLength%20(Decimal)%3A,as%20SliderMultiplier.<br/>
+    /// Note: Chosen to use double instead of decimal as the type, because maps like this exist:
+    /// https://osu.ppy.sh/beatmapsets/1219078#osu/2536330,
+    /// in which timing points can have beatLength be NaN, which cannot be represented as decimal.
+    /// </summary>
+    public double beatLength  { get; set; }
     public int meter  { get; set; }
     public int sampleSet  { get; set; }
     public int sampleIndex  { get; set; }
@@ -15,7 +21,7 @@ public class TimingPoint
 
     public TimingPoint() { }
 
-    public TimingPoint(int time, decimal beatLength, int meter, int sampleSet, int sampleIndex, int volume, bool uninherited, Effects effects)
+    public TimingPoint(int time, double beatLength, int meter, int sampleSet, int sampleIndex, int volume, bool uninherited, Effects effects)
     {
         this.time = time;
         this.beatLength = beatLength;
